+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { Article, Settings, StatusListener } from './commands';
 import { Template } from './commands/Template';
 import { TaxonomyType } from './models';
+import { TagType } from './viewpanel/TagType';
 import { ExplorerView } from './webview/ExplorerView';
 
 let frontMatterStatusBar: vscode.StatusBarItem;
@@ -18,12 +19,18 @@ export function activate({ subscriptions, extensionUri }: vscode.ExtensionContex
 		}
 	});
 
-	let insertTags = vscode.commands.registerCommand('frontMatter.insertTags', () => {
-		Article.insert(TaxonomyType.Tag);
+	let insertTags = vscode.commands.registerCommand('frontMatter.insertTags', async () => {
+		// Article.insert(TaxonomyType.Tag);
+		await vscode.commands.executeCommand('workbench.view.extension.frontmatter-explorer');
+		await vscode.commands.executeCommand('workbench.action.focusSideBar');
+		explorerSidebar.triggerInputFocus(TagType.tags);
 	});
 
-	let insertCategories = vscode.commands.registerCommand('frontMatter.insertCategories', () => {
-		Article.insert(TaxonomyType.Category);
+	let insertCategories = vscode.commands.registerCommand('frontMatter.insertCategories', async () => {
+		// Article.insert(TaxonomyType.Category);
+		await vscode.commands.executeCommand('workbench.view.extension.frontmatter-explorer');
+		await vscode.commands.executeCommand('workbench.action.focusSideBar');
+		explorerSidebar.triggerInputFocus(TagType.categories);
 	});
 
 	let createTag = vscode.commands.registerCommand('frontMatter.createTag', () => {