+++ src/webview/ExplorerView.ts
@@ -54,6 +54,8 @@ export class ExplorerView implements WebviewViewProvider, Disposable {
    */
   public async resolveWebviewView(webviewView: WebviewView, context: WebviewViewResolveContext, token: CancellationToken): Promise<void> {
 
+    console.log(context);
+
     this.panel = webviewView;
 
     webviewView.webview.options = {
@@ -122,6 +124,18 @@ export class ExplorerView implements WebviewViewProvider, Disposable {
    */
   public pushMetadata(metadata: any) {
     this.postWebviewMessage({ command: Command.metadata, data: metadata });
+  }
+
+  /**
+   * Allows the webview panel to focus on tags or categories input
+   * @param tagType 
+   */
+  public triggerInputFocus(tagType: TagType) {
+    if (tagType === TagType.tags) {
+      this.postWebviewMessage({ command: Command.focusOnTags });
+    } else {
+      this.postWebviewMessage({ command: Command.focusOnCategories });
+    }
   }
 
   /**
@@ -204,7 +218,7 @@ export class ExplorerView implements WebviewViewProvider, Disposable {
    * Post data to the panel
    * @param msg 
    */
-  private postWebviewMessage(msg: { command: Command, data: any }) {
+  private postWebviewMessage(msg: { command: Command, data?: any }) {
     this.panel!.webview.postMessage(msg);
   }
 
@@ -226,7 +240,7 @@ export class ExplorerView implements WebviewViewProvider, Disposable {
     const styleVSCodeUri = webView.asWebviewUri(Uri.joinPath(this.extPath, 'assets/media', 'vscode.css'));
     const styleResetUri = webView.asWebviewUri(Uri.joinPath(this.extPath, 'assets/media', 'reset.css'));
     const stylesUri = webView.asWebviewUri(Uri.joinPath(this.extPath, 'assets/media', 'styles.css'));
-    const scriptUri = webView.asWebviewUri(Uri.joinPath(this.extPath, 'dist', 'bundle.js'));
+    const scriptUri = webView.asWebviewUri(Uri.joinPath(this.extPath, 'dist', 'viewpanel.js'));
     const nonce = this.getNonce();
 
     return `